norm対応
